♻️ Refactor to making things easier to understand and avoid window issues
