+++ components/ImprovedSearch/index.ts
@@ -1,6 +1,6 @@
 export { SearchBox } from "./SearchBox"
 export { Hit } from "./Hit"
 export { Hits } from "./Hits"
-export { algoliaNextJsHistoryRouter } from "./algolia-nextjs-history-router"
+export { algoliaNextJsHistoryRouter } from "./lib/algolia-nextjs-history-router"
 
 export type { HitProps } from "./Hit"
